move `RECOMMENDED_MIN_FEE_VALUE` and `FEE_CONTRACT` to the common package
