+++ packages/common/src/constants.ts
@@ -1,0 +1,3 @@
+export const FEE_CONTRACT =
+  "1005040004000e36100204a00b08cd0279be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f81798ea02d192a39a8cc7a701730073011001020402d19683030193a38cc7b2a57300000193c2b2a57301007473027303830108cdeeac93b1a57304";
+export const RECOMMENDED_MIN_FEE_VALUE = BigInt(1100000);
