+++ packages/core/src/builder/pluginContext.spec.ts
@@ -3,7 +3,8 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 import { NotAllowedTokenBurning } from "../errors";
 import { OutputBuilder } from "./outputBuilder";
 import { createPluginContext, type FleetPluginContext } from "./pluginContext";
-import { RECOMMENDED_MIN_FEE_VALUE, TransactionBuilder } from "./transactionBuilder";
+import { TransactionBuilder } from "./transactionBuilder";
+import { RECOMMENDED_MIN_FEE_VALUE } from "@fleet-sdk/common";
 
 describe("Plugin context", () => {
   const creationHeight = 894169;
