+++ packages/core/src/builder/transactionBuilder.ts
@@ -10,11 +10,13 @@ import {
   byteSizeOf,
   chunk,
   ensureBigInt,
+  FEE_CONTRACT,
   first,
   isDefined,
   isHex,
   isUndefined,
   Network,
+  RECOMMENDED_MIN_FEE_VALUE,
   some,
   utxoDiff,
   utxoSum
@@ -45,10 +47,6 @@ export type ConfigureCallback = (settings: TransactionBuilderSettings) => void;
 export type SelectorCallback = (selector: SelectorSettings) => void;
 export type FleetPlugin = (context: FleetPluginContext) => void;
 export type CollectionLike<T> = { toArray(): T[] };
-
-export const RECOMMENDED_MIN_FEE_VALUE = BigInt(1100000);
-export const FEE_CONTRACT =
-  "1005040004000e36100204a00b08cd0279be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f81798ea02d192a39a8cc7a701730073011001020402d19683030193a38cc7b2a57300000193c2b2a57301007473027303830108cdeeac93b1a57304";
 
 type EjectorContext = {
   inputs: InputsCollection;
