+++ packages/core/src/index.ts
@@ -2,5 +2,10 @@ export * from "./builder";
 export * from "./models";
 export * from "./errors";
 export * from "./constantSerializer";
-export { AddressType, Network } from "@fleet-sdk/common";
+export {
+  AddressType,
+  Network,
+  RECOMMENDED_MIN_FEE_VALUE,
+  FEE_CONTRACT
+} from "@fleet-sdk/common";
 export type { Box, Amount, TokenAmount } from "@fleet-sdk/common";
